+++ scripts/discrete_representation/voxelization/voxelization_inverse.py
@@ -2,7 +2,7 @@
 Showcase of computing the inverse of a voxelization.
 """
 import volmdlr
-from volmdlr.core import VolumeModel
+from volmdlr.model import VolumeModel
 from volmdlr.discrete_representation import MatrixBasedVoxelization
 from volmdlr.primitives3d import Cylinder