+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## v0.18.0 [Future]
 ### New Features
 - new module: shapes.py
+- New module model.py to hold VolumeModel object.
 
 #### shapes.py
 - Shape: volume, bounding_box
@@ -15,8 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shape: add boolean operations: subtraction, union, intersection
 - Shape: to_dict, dict_to_object, to_brep, from_brep, from_brep_stream, to_brep_stream, 
 
+
 #### core.py
 - VolumeModel: from_step using OCP step reader
+- move VolumeModel from core.py to model.py.
 
 #### display.py
 - Mesh3D: rotation / translation / frame_mapping